+++ src/client/java/techreborn/client/compat/rei/ReiPlugin.java
@@ -60,10 +60,7 @@ import reborncore.client.gui.GuiBuilder;
 import reborncore.client.gui.GuiSprites;
 import reborncore.client.gui.config.GuiTab;
 import techreborn.TechReborn;
-import techreborn.client.compat.rei.fluidgenerator.FluidGeneratorRecipeCategory;
-import techreborn.client.compat.rei.fluidreplicator.FluidReplicatorRecipeCategory;
 import techreborn.client.compat.rei.machine.*;
-import techreborn.client.compat.rei.rollingmachine.RollingMachineCategory;
 import techreborn.init.ModRecipes;
 import techreborn.init.TRContent;
 import techreborn.init.TRContent.Machine;
