+++ src/client/java/techreborn/client/compat/rei/machine/RollingMachineCategory.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package techreborn.client.compat.rei.rollingmachine;
+package techreborn.client.compat.rei.machine;
 
 import me.shedaniel.rei.api.client.gui.Renderer;
 import me.shedaniel.rei.api.common.category.CategoryIdentifier;
