+++ src/client/java/techreborn/client/compat/rei/machine/SawmillCategory.java
@@ -32,7 +32,7 @@ import net.minecraft.recipe.RecipeType;
 import net.minecraft.text.Text;
 import reborncore.client.gui.GuiBuilder;
 import reborncore.common.crafting.RebornRecipe;
-import techreborn.client.compat.rei.MachineRecipeDisplay;
+import techreborn.compat.rei.MachineRecipeDisplay;
 import techreborn.client.compat.rei.ReiPlugin;
 
 import java.text.DecimalFormat;
