+++ src/main/java/techreborn/compat/rei/RollingMachineDisplay.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package techreborn.client.compat.rei.rollingmachine;
+package techreborn.compat.rei;
 
 import me.shedaniel.rei.api.common.category.CategoryIdentifier;
 import me.shedaniel.rei.plugin.common.displays.crafting.DefaultShapedDisplay;
